Added hero_type to the database
User can now choose their heroType. This will affect the visuals of the website and the contents of the /my-profile page. Later I will create a custom dropdown menu, to create the proper visuals. I will also need to add some validation.

Also improved the JWT system by allowing the refresh token to refresh itself as it refreshes the access token. This is done to avoid hard log-outs in the middle of a session.